Revert "Reuse http client"
