+++ lib/trade_tariff_frontend/request_forwarder.rb
@@ -24,7 +24,8 @@ module TradeTariffFrontend
                              .split('/')
                              .reject { |p| p.empty? || p == 'api' }
                              .first || "v#{Rails.configuration.x.backend.api_version}"
-        response = api.send(
+        conn = Faraday.new
+        response = conn.send(
           rackreq.request_method.downcase,
           request_url_for(rackreq)
         ) do |req|
@@ -96,10 +97,6 @@ module TradeTariffFrontend
       prefix = "/#{choice}"
 
       rackreq.path_info = rackreq.path_info.sub(prefix, '') if choice.present?
-    end
-
-    def api
-      TradeTariffFrontend::ServiceChooser.api_client_with_forwarding
     end
   end
 end
